Implement login page with redirect logic & OAuth2
