+++ app/routes/action.logout-user.ts
@@ -19,9 +19,11 @@ export const action = async ({ request }: ActionFunctionArgs) => {
   const formData = await request.formData();
   const logoutType = formData.get("logoutType") as LogoutType;
 
+  console.log(`Logging out user, type: ${logoutType}`);
+
   const session = await getSession(request.headers.get("Cookie"));
 
-  return redirect(`/?status=${logoutType}`, {
+  return redirect(`/login?status=${logoutType}`, {
     headers: {
       "Set-Cookie": await destroySession(session),
     },
